fix: One component per file for Time (#438)
* Rename file

* Rename temp

* Restore original

* Changes to refactor Time into separate source files

* Remove script file
