+++ src/TimePicker/TimePicker.js
@@ -1,6 +1,6 @@
 import { Popover } from '../Popover/Popover';
 import PropTypes from 'prop-types';
-import { Time } from '../Time/Time';
+import Time from '../Time/Time';
 import React, { Component } from 'react';
 import TimePickerItem, { CLOCK } from './_TimePickerItem';
